Fix compilation failure on Win64 since r14285.
R=dslomov@chromium.org

Review URL: https://codereview.chromium.org/14238012

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@14289 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/runtime.cc
@@ -829,7 +829,7 @@ RUNTIME_FUNCTION(MaybeObject*, Runtime_TypedArrayInitialize) {
       *isolate->factory()->NewNumber(static_cast<double>(length)));
   Handle<ExternalArray> elements =
       isolate->factory()->NewExternalArray(
-          length, arrayType,
+          static_cast<int>(length), arrayType,
           static_cast<uint8_t*>(buffer->backing_store()) + byte_offset);
   Handle<Map> map =
       isolate->factory()->GetElementsTransitionMap(holder, elementsKind);
